EDG and algo restructure.
