+++ atl-checker/Cargo.toml
@@ -22,6 +22,8 @@ use-counts = []
 test-env-log = { version = "0.2.5", features = ["trace"], default-features = false }
 criterion = "0.3"
 tracing-subscriber = "0.2.17"
+serde_json = "1.0.59"
+num_cpus = "1.13.0"
 
 [[bench]]
 name = "benchmark_solver"
