+++ atl-checker/src/algorithms/certain_zero/search_strategy/mod.rs
@@ -1,8 +1,7 @@
+use crate::atl::atl_cgs_edg::{Edge, NegationEdge, Vertex};
+
 pub mod bfs;
 pub mod dfs;
-
-use crate::common::{Edge, NegationEdge};
-use crate::edg::Vertex;
 
 /// A SearchStrategy defines in which order safe edges of an EDG is processed first in the
 /// certain zero algorithm.
