+++ atl-checker/src/atl/mod.rs
@@ -1,4 +1,4 @@
+pub mod atl_cgs_edg;
 pub mod common;
-pub mod dependencygraph;
 pub mod formula;
 pub mod gamestructure;
