+++ atl-checker/src/lib.rs
@@ -7,13 +7,9 @@ extern crate tracing;
 
 #[macro_use]
 mod simple_edg;
+pub mod algorithms;
 pub mod analyse;
 pub mod atl;
-mod com;
-pub mod common;
-pub mod edg;
 pub mod lcgs;
 #[cfg(feature = "graph-printer")]
 pub mod printer;
-pub mod search_strategy;
-pub mod solve_set;
